began working on order and made changes to bag

+++ pages/api/users/delete.js
@@ -1,13 +1,18 @@
 import User from "../../../models/User";
 import { dbConnect, dbDisconnect } from "../../../utils/database";
+import { verifyToken, verifyIsAdmin } from "../../../utils/auth";
 
-export default async function handler(req, res) {
+async function handler(req, res) {
+    if (req.method !== "DELETE") return res.status(405).json({ error: "DELETE method expected." });
+
     try {
         await dbConnect();
         await User.deleteMany({});
         await dbDisconnect();
-        res.status(200).json({ message: "Users deleted" });
+        return res.status(200).json({ message: "Users deleted" });
     } catch (err) {
-        res.status(500).json({ error: err.message });
+        return res.status(500).json({ error: err.message });
     }
 }
+
+export default verifyToken(verifyIsAdmin(handler));
